Give the fitweird div an ID
It's be nice to have an ID on the whole fitweird div so we can manipulate it / check for it / etc. without having to go deeper and check for a span with an ID.

This adds an id of 'fitweird' to the instance created by init();
It also changes the conditional to query for this ID rather than #fitweird-width-px. Seems cleaner.